feat: Refactor frontend to use ES modules
- Restructured the frontend JavaScript into modules for better organization and maintainability.
- Separated concerns like state management, rendering, WebSocket handling, and user interactions into their own files.
- Updated index.html to load the main script as a module.
