+++ package.json
@@ -7,6 +7,7 @@
     "axios": "^1.8.2",
     "dotenv": "^17.2.0",
     "express": "^4.21.2",
-    "express-ws": "^5.0.2"
+    "express-ws": "^5.0.2",
+    "helmet": "^8.1.0"
   }
 }
