+++ public/css/style.css
@@ -1,0 +1,103 @@
+body {
+    @apply bg-zinc-950 text-xl overflow-hidden h-full w-full fixed font-[Inter] text-zinc-100;
+}
+
+#connectionPopup {
+    @apply fixed top-5 left-1/2 transform -translate-x-1/2 z-50 bg-zinc-900 rounded-3xl p-4 shadow-xl transition-all duration-300 pointer-events-none;
+}
+
+#connectionPopup .icon {
+    @apply text-red-400 flex-shrink-0;
+}
+
+#connectionPopup .icon i {
+    @apply text-3xl;
+}
+
+#connectionPopup .content {
+    @apply flex-1 flex flex-col gap-0;
+}
+
+#connectionPopup h3 {
+    @apply text-sm font-semibold mb-1 m-0;
+}
+
+#connectionPopup p {
+    @apply text-xs text-zinc-300 m-0;
+}
+
+#reconnectBtn {
+    @apply bg-blue-600 hover:bg-blue-700 disabled:bg-zinc-600 disabled:cursor-not-allowed border-none rounded-lg px-3 py-2 text-base font-medium cursor-pointer flex items-center gap-1.5 transition-colors duration-200 whitespace-nowrap;
+}
+
+#reconnectIcon {
+    @apply text-4;
+}
+
+.top-left-controls {
+    @apply fixed top-2 left-2 flex gap-3 items-center z-20;
+}
+
+.top-left-controls label {
+    @apply sr-only;
+}
+
+#servers {
+    @apply bg-zinc-800 px-2 py-1 rounded text-base sm:text-xl font-[Inter];
+}
+
+#players {
+    @apply hidden;
+}
+
+.zoom-controls {
+    @apply fixed top-2 right-2 flex gap-2 sm:hidden z-20;
+}
+
+.zoom-controls button {
+    @apply bg-zinc-800 w-10 h-10 flex items-center justify-center rounded-full;
+}
+
+.zoom-controls button span {
+    @apply material-symbols-outlined;
+}
+
+.canvas-container {
+    @apply w-full h-full;
+}
+
+#tooltip {
+    @apply pl-3 pr-3.5 py-3 bg-zinc-800 rounded-2xl inline-flex flex-col justify-start items-start gap-2.5 fixed hidden sm:max-w-md max-w-[90vw] z-20;
+}
+
+#tooltip > div {
+    @apply inline-flex justify-center items-center gap-1.5;
+}
+
+#tooltip > div > span {
+    @apply material-symbols-outlined;
+}
+
+#tooltip > div > div {
+    @apply text-xl;
+}
+
+#pop-out {
+    @apply fixed bottom-2 right-12 flex p-2 align-center bg-zinc-800 rounded-lg z-20;
+}
+
+#pop-out i {
+    @apply material-symbols-outlined;
+}
+
+.github-link {
+    @apply fixed bottom-2 right-2 bg-zinc-800 p-2 rounded-lg hover:bg-zinc-700 transition z-30;
+}
+
+.github-link img {
+    @apply w-7 h-7;
+}
+
+#hover-hint {
+    @apply fixed bottom-2 left-3 z-20 text-zinc-500 text-sm hidden sm:block h-auto;
+}
