+++ public/js/constants.js
@@ -1,0 +1,38 @@
+export const WORLD_BOUNDS = {
+	TOP_LEFT: { x: -23818, y: -10426 },
+	BOTTOM_RIGHT: { x: 20504, y: 11377 },
+};
+
+export const STALE_SERVER_TIMEOUT = 30000;
+
+export const MAP_CONFIG = {
+	rows: 1,
+	cols: 16,
+	totalWidth: 28680,
+	totalHeight: 13724,
+};
+
+export const AREA_MARKERS = {
+	"Gleethrop End": { x: 1274, y: 3563 },
+	"Groenewoud": { x: -14658, y: -3762 },
+	"Dovedale East": { x: 1231, y: 534 },
+	"Fanory Mill": { x: -16821, y: -3954 },
+	"Mazewood": { x: -4650, y: 5798 },
+	"Conby": { x: -11688, y: -3270 },
+	"Codsall Castle": { x: 9991, y: 5236 },
+	"Masonfield": { x: 10667, y: -881 },
+	"Benyhone Loop": { x: -19532, y: -5201 },
+	"Perthtyne": { x: -490, y: 5268 },
+	"Ashburn": { x: -22012, y: -6729 },
+	"Cosdale Harbour": { x: 4325, y: -2518 },
+	"Glassbury Junction": { x: 11592, y: 8663 },
+	"Dovedale Central": { x: 3157, y: 805 },
+	"Wington Mount": { x: 2922, y: -2830 },
+	"Marigot Crossing": { x: 7692, y: 2205 },
+	"Satus": { x: -7485, y: -3055 },
+};
+
+export const COLORS = [
+	"#FD2943", "#01A2FF", "#02B857", "#A75EB8",
+	"#F58225", "#F5CD30", "#E8BAC8", "#D7C59A",
+];
